requirements.txt
pip
pysqlite3
discord-py-interactions
python-dotenv
pandas
rapidfuzz
